Add new member_of_collections display decorator mixin
