+++ app/decorators/ephemera_box_decorator.rb
@@ -1,24 +1,15 @@
 # frozen_string_literal: true
 class EphemeraBoxDecorator < Valkyrie::ResourceDecorator
+  include MemberOfCollectionsDisplay
   self.display_attributes = [
     :barcode,
     :box_number,
     :shipped_date,
     :tracking_number,
-    :member_of_collections,
     :visibility
   ]
 
   self.iiif_manifest_attributes = []
-
-  def member_of_collections
-    @member_of_collections ||=
-      begin
-        query_service.find_references_by(resource: model, property: :member_of_collection_ids)
-                     .map(&:decorate)
-                     .map(&:title).to_a
-      end
-  end
 
   def members
     @members ||= query_service.find_members(resource: model)
