+++ app/decorators/ephemera_folder_decorator.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 class EphemeraFolderDecorator < Valkyrie::ResourceDecorator
+  include MemberOfCollectionsDisplay
   self.display_attributes = [
     :barcode,
     :folder_number,
@@ -25,19 +26,10 @@ class EphemeraFolderDecorator < Valkyrie::ResourceDecorator
     :dspace_url,
     :source_url,
     :visibility
-  ]
+  ] - [:member_of_collections]
   self.iiif_manifest_attributes = display_attributes + [:title] - \
                                   imported_attributes(Schema::Common.attributes) - \
                                   Schema::IIIF.attributes - [:visibility, :internal_resource, :rights_statement, :rendered_rights_statement, :thumbnail_id]
-
-  def member_of_collections
-    @member_of_collections ||=
-      begin
-        query_service.find_references_by(resource: model, property: :member_of_collection_ids)
-                     .map(&:decorate)
-                     .map(&:title).to_a
-      end
-  end
 
   def members
     @members ||= query_service.find_members(resource: model)
