+++ app/decorators/scanned_map_decorator.rb
@@ -1,17 +1,9 @@
 # frozen_string_literal: false
 class ScannedMapDecorator < Valkyrie::ResourceDecorator
-  self.display_attributes += Schema::Geo.attributes + [:member_of_collections, :rendered_coverage] - [:thumbnail_id, :coverage, :cartographic_projection]
+  include MemberOfCollectionsDisplay
+  self.display_attributes += Schema::Geo.attributes + [:rendered_coverage] - [:thumbnail_id, :coverage, :cartographic_projection]
   self.iiif_manifest_attributes = display_attributes + [:title] - \
                                   Schema::IIIF.attributes - [:visibility, :internal_resource, :rights_statement, :rendered_rights_statement, :thumbnail_id]
-
-  def member_of_collections
-    @member_of_collections ||=
-      begin
-        query_service.find_references_by(resource: model, property: :member_of_collection_ids)
-                     .map(&:decorate)
-                     .map(&:title).to_a
-      end
-  end
 
   def members
     @members ||= query_service.find_members(resource: model)
