+++ app/decorators/scanned_resource_decorator.rb
@@ -1,19 +1,11 @@
 # frozen_string_literal: true
 class ScannedResourceDecorator < Valkyrie::ResourceDecorator
-  self.display_attributes += Schema::Common.attributes + imported_attributes(Schema::Common.attributes) + [:member_of_collections, :rendered_holding_location] - [:thumbnail_id]
+  include MemberOfCollectionsDisplay
+  self.display_attributes += Schema::Common.attributes + imported_attributes(Schema::Common.attributes) + [:rendered_holding_location] - [:thumbnail_id]
   self.iiif_manifest_attributes = display_attributes + [:title] - \
                                   imported_attributes(Schema::Common.attributes) - \
                                   Schema::IIIF.attributes - [:visibility, :internal_resource, :rights_statement, :rendered_rights_statement, :thumbnail_id]
   delegate(*Schema::Common.attributes, to: :primary_imported_metadata, prefix: :imported)
-
-  def member_of_collections
-    @member_of_collections ||=
-      begin
-        query_service.find_references_by(resource: model, property: :member_of_collection_ids)
-                     .map(&:decorate)
-                     .map(&:title).to_a
-      end
-  end
 
   def members
     @members ||= query_service.find_members(resource: model)
