+++ spec/decorators/ephemera_box_decorator_spec.rb
@@ -33,16 +33,4 @@ RSpec.describe EphemeraBoxDecorator do
       expect(resource.decorate.folders.to_a.first).to be_a EphemeraFolder
     end
   end
-  context 'within a collection' do
-    let(:collection) do
-      adapter = Valkyrie::MetadataAdapter.find(:indexing_persister)
-      res = FactoryGirl.build(:collection)
-      adapter.persister.save(resource: res)
-    end
-    let(:resource) { FactoryGirl.create_for_repository(:ephemera_box, member_of_collection_ids: [collection.id]) }
-    it 'retrieves the title of parents' do
-      expect(resource.decorate.member_of_collections.to_a).not_to be_empty
-      expect(resource.decorate.member_of_collections.to_a.first).to eq 'Title'
-    end
-  end
 end
