+++ spec/decorators/ephemera_folder_decorator_spec.rb
@@ -52,18 +52,6 @@ RSpec.describe EphemeraFolderDecorator do
       expect(resource.decorate.members.to_a.first).to be_a FileSet
     end
   end
-  context 'within a collection' do
-    let(:collection) do
-      adapter = Valkyrie::MetadataAdapter.find(:indexing_persister)
-      res = FactoryGirl.build(:collection)
-      adapter.persister.save(resource: res)
-    end
-    let(:resource) { FactoryGirl.create_for_repository(:ephemera_folder, member_of_collection_ids: [collection.id]) }
-    it 'retrieves the title of parents' do
-      expect(resource.decorate.member_of_collections.to_a).not_to be_empty
-      expect(resource.decorate.member_of_collections.to_a.first).to eq 'Title'
-    end
-  end
   context "within a box" do
     let(:resource) { FactoryGirl.create_for_repository(:ephemera_folder) }
     it "can return the box it's a member of" do
