+++ spec/decorators/scanned_resource_decorator_spec.rb
@@ -33,4 +33,16 @@ RSpec.describe ScannedResourceDecorator do
       end
     end
   end
+  context 'within a collection' do
+    let(:collection) do
+      adapter = Valkyrie::MetadataAdapter.find(:indexing_persister)
+      res = FactoryGirl.build(:collection)
+      adapter.persister.save(resource: res)
+    end
+    let(:resource) { FactoryGirl.create_for_repository(:scanned_resource, member_of_collection_ids: [collection.id]) }
+    it 'retrieves the title of parents' do
+      expect(resource.decorate.member_of_collections.to_a).not_to be_empty
+      expect(resource.decorate.member_of_collections.to_a.first).to eq 'Title'
+    end
+  end
 end
